requirements.txt
matplotlib==3.8.1
numpy==1.26.1
opencv_python==4.8.1.78
Pillow==10.0.1
Pillow==10.1.0
pycocotools==2.0.7
torch==2.1.0
torchvision==0.16.0